add calc.h utils.h
