+++ includes/utils.h
@@ -1,0 +1,19 @@
+
+#ifndef UTILS_H
+# define UTILS_H
+
+#if (__linux)
+	#include "mlx.h"
+#else
+	#include "mlx.h" 
+#endif
+#include "libft.h"
+#include "get_next_line.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdbool.h>
+#include <unistd.h>
+#include <fcntl.h>
+#include <math.h>
+
+#endif
