+++ srcs/calc_intersection.c
@@ -1,16 +1,5 @@
 #include "ogv1.h"
-
-double calc_distance(double ray_angle, t_player player, t_intersection res);
-double calc_offset(double ray_angle, enum e_axis axis, t_intersection inter, t_map map, t_mlx *mlx);
-int		get_vartical_first(t_mlx *mlx, int *pos_y, int ray_angle);
-int		get_horizontal_first(t_mlx *mlx, int *pos_y, double ray_angle);
-void	display_vartical_grid_intersection(t_mlx *mlx, t_intersection pos, int color, double ray_angle);
-void	display_horizontal_grid_intersection(t_mlx *mlx, t_intersection pos, int color, double ray_angle);
-t_intersection display_vertical_intersection(t_mlx *mlx, t_intersection pos, int color, double ray_angle);
-t_intersection display_horizontal_intersection(t_mlx *mlx, t_intersection pos, int color, double ray_angle);
-t_intersection	find_vartical_intersection(t_mlx *mlx, double ray_angle);
-t_intersection	find_horizontal_intersection(t_mlx *mlx, double ray_angle);
-t_intersection calc_intersection(t_mlx *mlx, double ray_angle);
+#include "calc.h"
 
 double calc_distance(double ray_angle, t_player player, t_intersection res)
 {
